+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps/Interop/Modules/AzSources.cs
@@ -47,6 +47,14 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Interop.Modules
         /// <param name="sourceIds">The collection of source IDs to clear.</param>
         ValueTask ClearById(string mapId, IEnumerable<string> sourceIds);
 
+        /// <summary>
+        /// Retrieves a shape from a specific source on the specified map by its ID.
+        /// </summary>
+        /// <param name="mapId"></param>
+        /// <param name="sourceId"></param>
+        /// <param name="shapeId"></param>
+        /// <returns></returns>
+        ValueTask<IJSObjectReference?> GetShapeById(string mapId, string sourceId, string shapeId);
 
         /// <summary>
         /// Retrieves the shapes from a specific source on the specified map.
@@ -140,6 +148,12 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Interop.Modules
         {
             var module = await moduleTask.Value;
             await module.InvokeVoidAsync(GetJsInteropMethod(), mapId, sourceIds);
+        }
+
+        public async ValueTask<IJSObjectReference?> GetShapeById(string mapId, string sourceId, string shapeId)
+        {
+            var module = await moduleTask.Value;
+            return await module.InvokeAsync<IJSObjectReference?>(GetJsInteropMethod(), mapId, sourceId, shapeId);
         }
 
         public async ValueTask<List<MapEventShape>> GetShapes(string mapId, string sourceId)